\name{NUEVOPAQUETE2-package}
\alias{NUEVOPAQUETE2-package}
\alias{NUEVOPAQUETE2}
\docType{package}
\title{
\packageTitle{NUEVOPAQUETE2}
}
\description{
\packageDescription{NUEVOPAQUETE2}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{NUEVOPAQUETE2}
\packageIndices{NUEVOPAQUETE2}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{NUEVOPAQUETE2}

Maintainer: \packageMaintainer{NUEVOPAQUETE2}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
